Added an option to fetch current version from milestones
Micrometer uses Nebula to retrieve the current version (don't ask my why). We don't have the `gradle.properties` file with the version number. 

This PR basing on the target branch (e.g. `1.2.x`) will resolve the latest open milestone starting with `major.minor` (e.g. `1.2.9`) and will assume that this is the project's version and the corresponding milestone. Then will continue with what the script does